refactor: extract preset management logic into dedicated PresetManager class
Introduce a new PresetManager class to centralize all preset-related operations, including category and tag management, UI updates, and event handling. This replaces scattered functions across multiple files and removes the deprecated initializePresetsEvents from global events. The change improves code organization, reduces coupling, and provides a single source of truth for preset state management.
